backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.33.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
